refactor(cli): split monolithic click app into subcommands
- Organize CLI commands into polylogue/cli/commands/
- Extract CLI helpers and formatting logic
- Update test_cli.py to support new structure
- Implement rich progress bar integration for 'run' command
